+++ test/test_invalid_headers.py
@@ -439,6 +439,71 @@ class TestFilter(object):
         if not (flags.is_trailer or flags.is_response_header)
     ]
 
+    invalid_request_header_blocks_bytes = (
+        # First, missing :method
+        (
+            (b':authority', b'google.com'),
+            (b':path', b'/'),
+            (b':scheme', b'https'),
+        ),
+        # Next, missing :path
+        (
+            (b':authority', b'google.com'),
+            (b':method', b'GET'),
+            (b':scheme', b'https'),
+        ),
+        # Next, missing :scheme
+        (
+            (b':authority', b'google.com'),
+            (b':method', b'GET'),
+            (b':path', b'/'),
+        ),
+        # Finally, path present but empty.
+        (
+            (b':authority', b'google.com'),
+            (b':method', b'GET'),
+            (b':scheme', b'https'),
+            (b':path', b''),
+        ),
+    )
+    invalid_request_header_blocks_unicode = (
+        # First, missing :method
+        (
+            (u':authority', u'google.com'),
+            (u':path', u'/'),
+            (u':scheme', u'https'),
+        ),
+        # Next, missing :path
+        (
+            (u':authority', u'google.com'),
+            (u':method', u'GET'),
+            (u':scheme', u'https'),
+        ),
+        # Next, missing :scheme
+        (
+            (u':authority', u'google.com'),
+            (u':method', u'GET'),
+            (u':path', u'/'),
+        ),
+        # Finally, path present but empty.
+        (
+            (u':authority', u'google.com'),
+            (u':method', u'GET'),
+            (u':scheme', u'https'),
+            (u':path', u''),
+        ),
+    )
+
+    # All headers that are forbidden from either request or response blocks.
+    forbidden_request_headers_bytes = (b':status',)
+    forbidden_request_headers_unicode = (u':status',)
+    forbidden_response_headers_bytes = (
+        b':path', b':scheme', b':authority', b':method'
+    )
+    forbidden_response_headers_unicode = (
+        u':path', u':scheme', u':authority', u':method'
+    )
+
     @pytest.mark.parametrize('validation_function', validation_functions)
     @pytest.mark.parametrize('hdr_validation_flags', hdr_validation_combos)
     @given(headers=HEADERS_STRATEGY)
@@ -489,6 +554,136 @@ class TestFilter(object):
     )
     def test_response_header_without_status(self, hdr_validation_flags):
         headers = [(b'content-length', b'42')]
+        with pytest.raises(h2.exceptions.ProtocolError):
+            list(h2.utilities.validate_headers(headers, hdr_validation_flags))
+
+    @pytest.mark.parametrize(
+        'hdr_validation_flags', hdr_validation_request_headers_no_trailer
+    )
+    @pytest.mark.parametrize(
+        'header_block',
+        (
+            invalid_request_header_blocks_bytes +
+            invalid_request_header_blocks_unicode
+        )
+    )
+    def test_outbound_req_header_missing_pseudo_headers(self,
+                                                        hdr_validation_flags,
+                                                        header_block):
+        with pytest.raises(h2.exceptions.ProtocolError):
+            list(
+                h2.utilities.validate_outbound_headers(
+                    header_block, hdr_validation_flags
+                )
+            )
+
+    @pytest.mark.parametrize(
+        'hdr_validation_flags', hdr_validation_request_headers_no_trailer
+    )
+    @pytest.mark.parametrize(
+        'header_block', invalid_request_header_blocks_bytes
+    )
+    def test_inbound_req_header_missing_pseudo_headers(self,
+                                                       hdr_validation_flags,
+                                                       header_block):
+        with pytest.raises(h2.exceptions.ProtocolError):
+            list(
+                h2.utilities.validate_headers(
+                    header_block, hdr_validation_flags
+                )
+            )
+
+    @pytest.mark.parametrize(
+        'hdr_validation_flags', hdr_validation_request_headers_no_trailer
+    )
+    @pytest.mark.parametrize(
+        'invalid_header',
+        forbidden_request_headers_bytes + forbidden_request_headers_unicode
+    )
+    def test_outbound_req_header_extra_pseudo_headers(self,
+                                                      hdr_validation_flags,
+                                                      invalid_header):
+        """
+        Outbound request header blocks containing the forbidden request headers
+        fail validation.
+        """
+        headers = [
+            (b':path', b'/'),
+            (b':scheme', b'https'),
+            (b':authority', b'google.com'),
+            (b':method', b'GET'),
+        ]
+        headers.append((invalid_header, b'some value'))
+        with pytest.raises(h2.exceptions.ProtocolError):
+            list(
+                h2.utilities.validate_outbound_headers(
+                    headers, hdr_validation_flags
+                )
+            )
+
+    @pytest.mark.parametrize(
+        'hdr_validation_flags', hdr_validation_request_headers_no_trailer
+    )
+    @pytest.mark.parametrize(
+        'invalid_header',
+        forbidden_request_headers_bytes
+    )
+    def test_inbound_req_header_extra_pseudo_headers(self,
+                                                     hdr_validation_flags,
+                                                     invalid_header):
+        """
+        Inbound request header blocks containing the forbidden request headers
+        fail validation.
+        """
+        headers = [
+            (b':path', b'/'),
+            (b':scheme', b'https'),
+            (b':authority', b'google.com'),
+            (b':method', b'GET'),
+        ]
+        headers.append((invalid_header, b'some value'))
+        with pytest.raises(h2.exceptions.ProtocolError):
+            list(h2.utilities.validate_headers(headers, hdr_validation_flags))
+
+    @pytest.mark.parametrize(
+        'hdr_validation_flags', hdr_validation_response_headers
+    )
+    @pytest.mark.parametrize(
+        'invalid_header',
+        forbidden_response_headers_bytes + forbidden_response_headers_unicode
+    )
+    def test_outbound_resp_header_extra_pseudo_headers(self,
+                                                       hdr_validation_flags,
+                                                       invalid_header):
+        """
+        Outbound response header blocks containing the forbidden response
+        headers fail validation.
+        """
+        headers = [(b':status', b'200')]
+        headers.append((invalid_header, b'some value'))
+        with pytest.raises(h2.exceptions.ProtocolError):
+            list(
+                h2.utilities.validate_outbound_headers(
+                    headers, hdr_validation_flags
+                )
+            )
+
+    @pytest.mark.parametrize(
+        'hdr_validation_flags', hdr_validation_response_headers
+    )
+    @pytest.mark.parametrize(
+        'invalid_header',
+        forbidden_response_headers_bytes
+    )
+    def test_inbound_resp_header_extra_pseudo_headers(self,
+                                                      hdr_validation_flags,
+                                                      invalid_header):
+        """
+        Inbound response header blocks containing the forbidden response
+        headers fail validation.
+        """
+        headers = [(b':status', b'200')]
+        headers.append((invalid_header, b'some value'))
         with pytest.raises(h2.exceptions.ProtocolError):
             list(h2.utilities.validate_headers(headers, hdr_validation_flags))